+++ sdk/storage/azure-storage-files-shares/src/share_service_client.cpp
@@ -103,7 +103,7 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
   }
 
   Azure::Response<Models::SetServicePropertiesResult> ShareServiceClient::SetProperties(
-      Models::FileServiceProperties properties,
+      Models::ShareFileServiceProperties properties,
       const SetServicePropertiesOptions& options,
       const Azure::Core::Context& context) const
   {
@@ -114,7 +114,7 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
         m_serviceUrl, *m_pipeline, context, protocolLayerOptions);
   }
 
-  Azure::Response<Models::FileServiceProperties> ShareServiceClient::GetProperties(
+  Azure::Response<Models::ShareFileServiceProperties> ShareServiceClient::GetProperties(
       const GetServicePropertiesOptions& options,
       const Azure::Core::Context& context) const
   {
@@ -122,12 +122,12 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
     auto protocolLayerOptions = _detail::ShareRestClient::Service::GetPropertiesOptions();
     auto result = _detail::ShareRestClient::Service::GetProperties(
         m_serviceUrl, *m_pipeline, context, protocolLayerOptions);
-    Models::FileServiceProperties ret;
+    Models::ShareFileServiceProperties ret;
     ret.Cors = std::move(result->Cors);
     ret.HourMetrics = std::move(result->HourMetrics);
     ret.MinuteMetrics = std::move(result->MinuteMetrics);
     ret.Protocol = std::move(result->Protocol);
-    return Azure::Response<Models::FileServiceProperties>(
+    return Azure::Response<Models::ShareFileServiceProperties>(
         std::move(ret), result.ExtractRawResponse());
   }