+++ sdk/storage/azure-storage-files-shares/CHANGELOG.md
@@ -14,10 +14,9 @@
 - Renamed `GetSharePropertiesResult` to `ShareProperties`.
 - Renamed `GetShareStatisticsResult` to `ShareStatistics`.
 - Renamed `GetShareDirectoryPropertiesResult` to `ShareDirectoryProperties`.
-- Renamed `GetShareFilePropertiesResult` to `ShareFileProperties`.
-- Renamed `GetShareFileRangeListResult` to `ShareFileRangeInfo`.
+- Renamed `GetShareFilePropertiesResult` to `ShareFileServiceProperties`.
 - `ShareClient::GetPermission` now returns `Azure::Response<std::string>`.
-- `ShareServiceClient::GetProperties` now returns `Azure::Response<Models::FileServiceProperties>`.
+- `ShareServiceClient::GetProperties` now returns `Azure::Response<Models::ShareFileServiceProperties>`.
 
 ## 12.0.0-beta.8 (2021-02-12)